+++ integration-tests/testsetups/ocr.go
@@ -4,7 +4,6 @@ package testsetups
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"math/big"
 	"math/rand"
 	"os"
@@ -136,13 +135,14 @@ func (o *OCRSoakTest) DeployEnvironment(customChainlinkNetworkTOML string) {
 		Test:            o.t,
 	}
 
-	cd, err := chainlink.NewDeployment(6, map[string]any{
-		"toml": client.AddNetworkDetailedConfig(config.BaseOCRP2PV1Config, customChainlinkNetworkTOML, network),
+	cd := chainlink.New(0, map[string]any{
+		"replicas": 6,
+		"toml":     client.AddNetworkDetailedConfig(config.BaseOCRP2PV1Config, customChainlinkNetworkTOML, network),
 		"db": map[string]any{
 			"stateful": true, // stateful DB by default for soak tests
 		},
 	})
-	require.NoError(o.t, err, "Error creating chainlink deployment")
+
 	testEnvironment := environment.New(baseEnvironmentConfig).
 		AddHelm(mockservercfg.New(nil)).
 		AddHelm(mockserver.New(nil)).
@@ -151,8 +151,8 @@ func (o *OCRSoakTest) DeployEnvironment(customChainlinkNetworkTOML string) {
 			Simulated:   network.Simulated,
 			WsURLs:      network.URLs,
 		})).
-		AddHelmCharts(cd)
-	err = testEnvironment.Run()
+		AddHelm(cd)
+	err := testEnvironment.Run()
 	require.NoError(o.t, err, "Error launching test environment")
 	o.testEnvironment = testEnvironment
 	o.namespace = testEnvironment.Cfg.Namespace
@@ -358,7 +358,7 @@ func (o *OCRSoakTest) LoadState() error {
 	}
 
 	testState := &OCRSoakTestState{}
-	saveData, err := ioutil.ReadFile(saveFileLocation)
+	saveData, err := os.ReadFile(saveFileLocation)
 	if err != nil {
 		return err
 	}
@@ -555,14 +555,13 @@ func (o *OCRSoakTest) setFilterQuery() {
 func (o *OCRSoakTest) observeOCREvents() error {
 	eventLogs := make(chan types.Log)
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
 	eventSub, err := o.chainClient.SubscribeFilterLogs(ctx, o.filterQuery, eventLogs)
+	cancel()
 	if err != nil {
 		return err
 	}
 
 	go func() {
-		defer cancel()
 		for {
 			select {
 			case event := <-eventLogs:
@@ -589,6 +588,7 @@ func (o *OCRSoakTest) observeOCREvents() error {
 						Msg("Error while subscribed to OCR Logs. Resubscribing")
 					ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
 					eventSub, err = o.chainClient.SubscribeFilterLogs(ctx, o.filterQuery, eventLogs)
+					cancel()
 				}
 			}
 		}